requirements.txt
discord.py==2.3.2
firebase-admin==6.2.0
python-dotenv==1.0.0
aiohttp==3.8.5
anthropic==0.7.7 